console: Refactor application redux primitives
